requirements.txt
streamlit==1.43.0
openai==1.37.0
python-dotenv==1.0.1
tqdm==4.66.4
stqdm==0.0.5
streamlit-clipboard==0.0.6
faiss-cpu==1.8.0.post1
numpy
tabulate
httpx[http2]==0.27.0
